+++ LifeLike.Frontend/src/app/modules/admin/pages/videos/videos.component.ts
@@ -3,6 +3,7 @@ import { MatDialog, MatDialogConfig } from '@angular/material';
 import { map } from 'rxjs/operators';
 import Video from '../../../../modules/video/models/Video';
 import Photo from '../../../photo/models/Photo';
+import { ConfirmDeleteComponent } from '../../dialogs/confirm-delete/confirm-delete.component';
 import { VideoCreateComponent } from '../../dialogs/videos-create/video-create.component';
 import { VideoEditComponent } from '../../dialogs/videos-edit/video-edit.component';
 import { AdminRestService } from '../../services/admin-rest.service';
@@ -21,23 +22,28 @@ export class VideosComponent implements OnInit {
               private readonly dialog: MatDialog) { }
 
   Remove(video: Video): void {
-    console.log('Remove');
-    console.log(video);
-    this.IsLoading = true;
-    this.restService.removeVideo(video.Id)
-      .subscribe(
-        () => {
-          this.IsLoading = false;
-          this.GetVideos();
-        },
-        (error) => {
-          this.error = error;
-          this.IsLoading = false;
-        });
+    const dialogConfig = new MatDialogConfig();
+    dialogConfig.data = video.Name;
+    dialogConfig.autoFocus = true;
+    const dialogRef = this.dialog.open(ConfirmDeleteComponent, dialogConfig);
+    dialogRef.afterClosed().subscribe((result) => {
+      if (result) {
+        this.IsLoading = true;
+        this.restService.removeVideo(video.Id)
+          .subscribe(
+            () => {
+              this.IsLoading = false;
+              this.GetVideos();
+            },
+            (error) => {
+              this.error = error;
+              this.IsLoading = false;
+            });
+      }
+    });
+
   }
   Edit(model: Video): void {
-    console.log('Edit');
-    console.log(model);
     const dialogConfig = new MatDialogConfig();
     dialogConfig.data = model;
     dialogConfig.disableClose = true;
@@ -48,7 +54,6 @@ export class VideosComponent implements OnInit {
     });
   }
   Create(): void {
-    console.log('Create');
     const dialogConfig = new MatDialogConfig();
     dialogConfig.disableClose = true;
     dialogConfig.autoFocus = true;
@@ -64,8 +69,6 @@ export class VideosComponent implements OnInit {
       .pipe(
         map((data: Video[]) => {
           this.IsLoading = false;
-          console.log(data);
-
           return data;
         }))
       .subscribe((p: Video[]) => this.Videos = p);
